Moved readyStorage to main.
