+++ tsdb/db.go
@@ -1455,88 +1455,3 @@ func exponential(d, min, max time.Duration) time.Duration {
 	}
 	return d
 }
-
-// ReadyStorage implements the Storage interface while allowing to set the actual
-// storage at a later point in time.
-type ReadyStorage struct {
-	mtx             sync.RWMutex
-	db              *DB
-	startTimeMargin int64
-}
-
-// Set the storage.
-func (s *ReadyStorage) Set(db *DB, startTimeMargin int64) {
-	s.mtx.Lock()
-	defer s.mtx.Unlock()
-
-	s.db = db
-	s.startTimeMargin = startTimeMargin
-}
-
-// Get the storage.
-func (s *ReadyStorage) Get() *DB {
-	if x := s.get(); x != nil {
-		return x
-	}
-	return nil
-}
-
-func (s *ReadyStorage) get() *DB {
-	s.mtx.RLock()
-	x := s.db
-	s.mtx.RUnlock()
-	return x
-}
-
-// StartTime implements the Storage interface.
-func (s *ReadyStorage) StartTime() (int64, error) {
-	if x := s.get(); x != nil {
-		var startTime int64
-
-		if len(x.Blocks()) > 0 {
-			startTime = x.Blocks()[0].Meta().MinTime
-		} else {
-			startTime = time.Now().Unix() * 1000
-		}
-		// Add a safety margin as it may take a few minutes for everything to spin up.
-		return startTime + s.startTimeMargin, nil
-	}
-
-	return math.MaxInt64, ErrNotReady
-}
-
-// Querier implements the Storage interface.
-func (s *ReadyStorage) Querier(ctx context.Context, mint, maxt int64) (storage.Querier, error) {
-	if x := s.get(); x != nil {
-		return x.Querier(ctx, mint, maxt)
-	}
-	return nil, ErrNotReady
-}
-
-// Appender implements the Storage interface.
-func (s *ReadyStorage) Appender() storage.Appender {
-	if x := s.get(); x != nil {
-		return x.Appender()
-	}
-	return notReadyAppender{}
-}
-
-type notReadyAppender struct{}
-
-func (n notReadyAppender) Add(l labels.Labels, t int64, v float64) (uint64, error) {
-	return 0, ErrNotReady
-}
-
-func (n notReadyAppender) AddFast(ref uint64, t int64, v float64) error { return ErrNotReady }
-
-func (n notReadyAppender) Commit() error { return ErrNotReady }
-
-func (n notReadyAppender) Rollback() error { return ErrNotReady }
-
-// Close implements the Storage interface.
-func (s *ReadyStorage) Close() error {
-	if x := s.Get(); x != nil {
-		return x.Close()
-	}
-	return nil
-}
